requirements.txt
cffi==1.14.1
cloudpickle==0.5.3
cycler==0.10.0
dask==0.18.2
decorator==4.3.0
imageio
kiwisolver==1.0.1
matplotlib
networkx
numpy
pandas
Pillow
pycparser==2.18
pygit
pyparsing==2.2.0
python-dateutil
pytz
PyWavelets
PyYAML==5.3.1
scikit-image
scikit-learn
scipy
six==1.15.0
toolz==0.9.0
torch
torchvision
tqdm==4.24.0
